add.1 jpa2 真是好难...

+++ src/main/java/com/spring/repository/AreaRepository.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
  * @version Created by zhenghuan on 2017/2/14
  */
 @Repository
-public interface AreaRepository extends JpaRepository<Area, Integer> {
+public interface AreaRepository extends JpaRepository<Area, Long> {
 }
